feat(ArtistPromoters): implement PromoterCard component and useArtistPromoters hook for improved promoter management
